pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agenteIA-ordenarTarea"
version = "0.1"
description = "Un agente de IA especializado en ordenar tareas por prioridad, tiempo y fecha fin"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [

]

[project.scripts]
agenteIA-ordenarTarea = "agenteIA_OrdenarTarea.main:run"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100